+++ packages/client/src/components/user-card-mini.vue
@@ -21,7 +21,7 @@ const props = defineProps<{
 
 const chart = $ref(null);
 
-os.api('charts/user/notes', { userId: props.user.id, limit: 16, span: 'day' }).then(res => {
+os.apiGet('charts/user/notes', { userId: props.user.id, limit: 16, span: 'day' }).then(res => {
 	chart = res;
 });
 </script>